Refactor texture save boundaries
