+++ src/app/mods/texture_save/__init__.py
@@ -1,0 +1,5 @@
+"""Save to Texture service boundary."""
+
+from .service import TextureSaveError, save_texture_color
+
+__all__ = ["TextureSaveError", "save_texture_color"]
