+++ src/core/textures/__init__.py
@@ -23,9 +23,8 @@ from .uv_coverage import (
     dilate_pixel_mask, rasterize_uv_coverage,
 )
 from .color_adjustment import (
-    COLOR_DEFAULTS, COLOR_RANGES, adjust_rgba_bytes,
-    apply_color_adjustment, apply_prepared_color_adjustment,
-    apply_prepared_color_u8, is_neutral_color_adjustment,
+    apply_prepared_color_adjustment, apply_prepared_color_u8,
+    is_neutral_color_adjustment,
     normalize_color_adjustment, prepare_color_adjustment,
     PreparedColorAdjustment,
 )
@@ -52,10 +51,6 @@ __all__ = [
     "collapse_pixel_mask_to_units",
     "dilate_pixel_mask",
     "rasterize_uv_coverage",
-    "COLOR_DEFAULTS",
-    "COLOR_RANGES",
-    "adjust_rgba_bytes",
-    "apply_color_adjustment",
     "apply_prepared_color_adjustment",
     "apply_prepared_color_u8",
     "is_neutral_color_adjustment",
